fix(test): rename latLng to latlng
