+++ test/models/ui/map_pin_details_test.dart
@@ -2,7 +2,7 @@ import "package:flutter_test/flutter_test.dart";
 import "package:google_maps_flutter/google_maps_flutter.dart";
 import "package:proxima/models/ui/map_pin_details.dart";
 
-import "../../mocks/data/latLng.dart";
+import "../../mocks/data/latlng.dart";
 
 void main() {
   group("Map pin testing", () {
